Migrate project to internationalized routing with locale-based pages
