+++ app/[locale]/about/page.tsx
