+++ app/[locale]/privacy/page.tsx
