--- app/test-page.tsx
@@ -1,8 +1,0 @@
-export default function TestPage() {
-  return (
-    <div>
-      <h1>Test Page</h1>
-      <p>This is a test page to check JSX syntax.</p>
-    </div>
-  )
-}
